# openapi.yaml
openapi: 3.0.3
info:
  title: UsahaKita API
  description: |
    Multi-tenant micro ERP API untuk operasi material, resep, produksi, produk jadi, penjualan, dan AI insight.

    ## Authentication
    Semua endpoint yang memerlukan autentikasi menggunakan cookie `session_token` (HttpOnly, Secure, SameSite=Strict).

    ## Tenant Isolation
    `tenant_id` selalu berasal dari session yang terotentikasi, bukan dari body request.

    ## Response Envelope
    Semua response menggunakan format:
    - Success: `{ "success": true, "data": ..., "meta": ... }`
    - Error: `{ "success": false, "error": { "code": "...", "message": "...", "fields": {} } }`
  version: 1.0.0
  contact:
    name: UsahaKita
  license:
    name: MIT
servers:
  - url: /api/v1
    description: API v1
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Registrasi akun dan usaha baru
      description: Membuat user baru, tenant baru, dan mengatur role OWNER. Mengembalikan session cookie.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: owner@tokomaju.com
              password: StrongPassword123!
              business_name: Toko Maju
              business_description: Usaha makanan rumahan
      responses:
        '201':
          description: Registrasi berhasil
          headers:
            Set-Cookie:
              schema:
                type: string
              example: 'session_token=abc123; Path=/; HttpOnly; Secure; SameSite=Strict; Max-Age=604800'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '409':
          $ref: '#/components/responses/EmailAlreadyExists'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/login:
    post:
      tags: [Auth]
      summary: Masuk ke akun
      description: Autentikasi user dengan email dan password. Mengembalikan session cookie.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: owner@tokomaju.com
              password: StrongPassword123!
      responses:
        '200':
          description: Login berhasil
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '401':
          $ref: '#/components/responses/InvalidCredentials'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/me:
    get:
      tags: [Auth]
      summary: Ambil data user dan bisnis saat ini
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '401':
          $ref: '#/components/responses/AuthRequired'
        '404':
          $ref: '#/components/responses/NotFound'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Keluar dari sesi
      description: Mencabut session di server dan menghapus cookie.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Logout berhasil
          headers:
            Set-Cookie:
              schema:
                type: string
              example: 'session_token=; Path=/; HttpOnly; Secure; SameSite=Strict; Max-Age=0'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NullSuccessResponse'
        '401':
          $ref: '#/components/responses/AuthRequired'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: header
      name: Cookie
      description: |
        Cookie `session_token` yang di-set saat login/register.
        Format: `Cookie: session_token=<token>`
...
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, business_name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        business_name:
          type: string
        business_description:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            business:
              $ref: '#/components/schemas/Business'
    User:
      type: object
      properties:
        id:
          type: string
          example: usr_01
        email:
          type: string
          format: email
          example: owner@tokomaju.com
    Business:
      type: object
      properties:
        id:
          type: string
          example: ten_01
        name:
          type: string
          example: Toko Maju
        description:
          type: string
          example: Usaha produksi dan penjualan makanan rumahan
...
    NullSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          nullable: true
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            fields:
              type: object
              additionalProperties:
                type: string
  responses:
    AuthRequired:
      description: Autentikasi diperlukan
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: AUTH_REQUIRED
              message: Authentication required
    InvalidCredentials:
      description: Email atau password salah
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: INVALID_CREDENTIALS
              message: Invalid email or password
    EmailAlreadyExists:
      description: Email sudah terdaftar
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: EMAIL_ALREADY_EXISTS
              message: Email already registered
    ValidationError:
      description: Validasi input gagal
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Validation failed
              fields:
                email: Harus diisi
    NotFound:
      description: Resource tidak ditemukan
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: RESOURCE_NOT_FOUND
              message: Resource not found